requirements.txt
discord.py
aiosqlite
colorama
os
asyncio
time
random
json